pkg
pkg is a simple package manager.  it supports installing, removing and searching of packages.  it supports simple dependency resolution (have no idea if it can handle multiple levels of dependencies... I assume the logic is good?)

packages are to be hosted on a simple web server.  they reside in /var/cache/packages/ and everything is tracked in /var/cache/packages/package.db.

I need to write a the logic to automatically wget or rsync the packages from a repo server.

I also need to write a repo list.  something like 'pkg-repo' -a http://1.2.3.4.  the repo list is basically just a SQL file that populates /var/cache/packages/package.db